ct->code->dev-ing: 继续开发用户部分

+++ ct-app/src/main/java/com/xjd/ct/app/biz/IdolBiz.java
@@ -31,7 +31,7 @@ public class IdolBiz {
 	 * @param limit
 	 * @return
 	 */
-	public List<UserForOtherVo> listIdols(Long userId, int offset, int limit) {
+	public List<UserForOtherVo> listIdols(Long userId, long offset, int limit) {
 		List<UserIdolModel> userIdolModelList = idolBizDao.selectUserIdolByUserIdAndPage(userId, offset, limit);
 
 		List<UserForOtherVo> userForOtherVoList = new ArrayList<UserForOtherVo>(userIdolModelList.size());
